requirements.txt
Django==5.0.3
djangorestframework==3.16.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.11
gunicorn==25.1.0
django-environ==0.11.2
pillow==11.0.0
celery==5.4.0
redis==5.0.1
django-filter==24.1
djangorestframework-simplejwt==5.3.2
drf-spectacular==0.27.0
